Added functionality in the load system.
Added a check to allow jQuery to be included in the DOM using @defer.
If the browser supports the addEventListener(), it supports the DOMContentLoaded.
gecko, webkit and IE10 follow the spec as is. <- That's to whom this is for, mostly
IE 9 shows "interactive" before the DOMContentLoaded event so it needs to be excluded from this.
Presto ignores defer and shows complete when it should show 'interactive' so it's already been taken care of.
IE 7 and IE8 won't even try to execute this because they don't support addEventListener()

If you have any problems with this code change just warn me. I can do some modifications if it's in my grasp.